refactor(dictation): extract dictation/pipeline.rs from commands/dictation (#597)
Item 9 of the architecture-cleanup roadmap. Pure refactor — no
behaviour change. Same shape as items 5/6 (split-by-concern peer
files under a directory module).

`commands/dictation.rs` was 1291 lines mixing thin command shells
(`audio_list_sources`, `start_dictation`, `stop_dictation`) with
~200 lines of orchestration helpers and ~660 lines of tests. The
helpers have nothing to do with the IPC surface — they're the
"do the thing" pipeline that `start_dictation` and `stop_dictation`
call into.

Changes:
- `commands/dictation.rs` -> `commands/dictation/mod.rs` (git mv;
  the module path stays `commands::dictation::*` so all `lib.rs`
  registrations continue to resolve unchanged).
- New `commands/dictation/pipeline.rs` (243 lines) with the
  extracted helpers, all `pub(super)`:
    - `start_dictation_inner` (the start orchestration body)
    - `strip_whisper_brackets`
    - `stop_audio_capture`
    - `load_vocabulary_prompt` / `load_replacement_rules`
    - `take_foreground_snapshot`
    - `write_to_clipboard`
    - `fire_ready_notification`
    - `spawn_history_create`
    - `capture_foreground`
- `mod.rs` re-imports them via `use pipeline::{...}` so the
  existing call sites in `start_dictation` / `stop_dictation`
  continue to read `helper(...)` rather than `pipeline::helper(...)`.

What stays in mod.rs: the three `#[tauri::command]` shells
(`audio_list_sources`, `start_dictation`, `stop_dictation`) plus
the existing test module. `stop_dictation` is still the heaviest
thing in the file — restructuring its body was considered but
rejected as too risky for the dictation hot path given autonomous-
mode discipline; deferred for a focused future PR.

Final shape:
- mod.rs: 1081 lines (was 1291; -16%, -210 LOC of helpers extracted)
- pipeline.rs: 243 lines (the extracted orchestration)

Visibility note: helpers are `pub(super)` — visible to the parent
`dictation` module and its `tests` module, not exposed outside the
`commands::dictation` namespace. The doc-comment header in
pipeline.rs explains why this is the right level: it's what makes
the "command shell vs pipeline" split structurally meaningful
rather than just naming.

Verified locally:
- cargo check --all-targets clean
- cargo clippy --all-targets -- -D warnings clean
- cargo clippy --lib --no-default-features -- -D warnings clean
- cargo fmt --all (applied)
- cargo test --lib: 412 pass, 0 fail (same count as before)
